// client/src/app/apis/model/nodeItem.ts
/**
 * api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v2.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { NodeReference } from './nodeReference';
import { NodeTag } from './nodeTag';
import { NodeUser } from './nodeUser';


export interface NodeItem { 
    id?: string;
    reference?: NodeReference;
    created?: Date;
    modified?: Date;
    name?: string;
    hasSolution?: boolean;
    solutionToId?: string;
    solutionTo?: string;
    path?: string;
    type?: NodeItem.TypeEnum;
    user?: NodeUser;
    content?: string;
    shared?: boolean;
    tags?: Array<NodeTag>;
}
export namespace NodeItem {
    export type TypeEnum = 'None' | 'Folder' | 'Blog' | 'Reference';
    export const TypeEnum = {
        None: 'None' as TypeEnum,
        Folder: 'Folder' as TypeEnum,
        Blog: 'Blog' as TypeEnum,
        Reference: 'Reference' as TypeEnum
    };
}