// client/src/app/apis/model/nodeTag.ts
/**
 * api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v2.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface NodeTag { 
    id?: string;
    name?: string;
    type?: NodeTag.TypeEnum;
    value?: string;
    values?: string;
}
export namespace NodeTag {
    export type TypeEnum = 'None' | 'Bool' | 'String' | 'Number' | 'Enum' | 'Url' | 'Private';
    export const TypeEnum = {
        None: 'None' as TypeEnum,
        Bool: 'Bool' as TypeEnum,
        String: 'String' as TypeEnum,
        Number: 'Number' as TypeEnum,
        Enum: 'Enum' as TypeEnum,
        Url: 'Url' as TypeEnum,
        Private: 'Private' as TypeEnum
    };
}